Wrong handling of JS-strings for alert
There is no need to create HTML-entities because we use UTF-8 charset and it is moreover wrong to create entities for JS alerts. The popup displays the entity wrongly as encoded instead of decoding it before like the browser does regularly. So e.g. german umlauts would be displayed weird/wrong.